a class Neuron that defines a single neuron performing binary classification (Based on 6-neuron.py):
Update the public method train to def train(self, X, Y, iterations=5000, alpha=0.05, verbose=True, graph=True, step=100):
Trains the neuron by updating the private attributes __W, __b, and __A
X is a numpy.ndarray with shape (nx, m) that contains the input data
nx is the number of input features to the neuron
m is the number of examples
Y is a numpy.ndarray with shape (1, m) that contains the correct labels for the input data
iterations is the number of iterations to train over
if iterations is not an integer, raise a TypeError with the exception iterations must be an integer
if iterations is not positive, raise a ValueError with the exception iterations must be a positive integer
alpha is the learning rate
if alpha is not a float, raise a TypeError with the exception alpha must be a float
if alpha is not positive, raise a ValueError with the exception alpha must be positive
verbose is a boolean that defines whether or not to print information about the training. If True, print Cost after {iteration} iterations: {cost} every step iterations:
Include data from the 0th and last iteration
graph is a boolean that defines whether or not to graph information about the training once the training has completed. If True:
Plot the training data every step iterations as a blue line
Label the x-axis as iteration
Label the y-axis as cost
Title the plot Training Cost
Include data from the 0th and last iteration
Only if either verbose or graph are True:
if step is not an integer, raise a TypeError with the exception step must be an integer
if step is not positive or is greater than iterations, raise a ValueError with the exception step must be positive and <= iterations
All exceptions should be raised in the order listed above
The 0th iteration should represent the state of the neuron before any training has occurred
You are allowed to use one loop
You can use import matplotlib.pyplot as plt
Returns: the evaluation of the training data after iterations of training have occurred