the class DeepNeuralNetwork to allow different activation functions (based on 27-deep_neural_network.py):
Update the __init__ method to def __init__(self, nx, layers, activation='sig'):
activation represents the type of activation function used in the hidden layers
sig represents a sigmoid activation
tanh represents a tanh activation
if activation is not sig or tanh, raise a ValueError with the exception: activation must be 'sig' or 'tanh'
Create the private attribute __activation and set it to the value of activation
Create a getter for the private attribute __activation
Update the forward_prop and gradient_descent instance methods to use the __activation function in the hidden layers
Because the training process takes such a long time, I have pre-trained a model for you to load and finish training (28-saved.pkl). This model has already been trained for 2000 iterations.